requirements.txt
keyring
requests
fabric
ruamel.yaml
bs4
jinja2
pre-commit
python-dateutil
isort
black
tqdm
humanize
lxml
